requirements.txt
requests
face_recognition
python-telegram-bot
psycopg2
sqlalchemy
ffmpeg
soundfile
